ai-engine/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
PyMuPDF
